server/requirements.txt
Django
djangorestframework
gunicorn

joblib
pandas
numpy 

scikit-learn